"El nombre del binario no es el acordado en el manpage", "La autenticacion con user/pass no funciona", "Un memory leak al matar el proceso"

+++ src/buffer.c
@@ -81,7 +81,7 @@ buffer_read(buffer *b) {
 
 inline bool 
 buffer_read_bytes(buffer *b, uint8_t dest[], size_t bytes) {
-    for(int i = 0; i < bytes; i++) {
+    for(size_t i = 0; i < bytes; i++) {
         if(!buffer_can_read(b)) return false;
         dest[i] = *b->read;
         buffer_read_adv(b, 1);
